refactor(tangle-dapp): Move transaction events to shared lib
