+++ apps/tangle-dapp/src/pages/restake/withdraw/WithdrawRequestTableActions.tsx
@@ -1,10 +1,9 @@
+import { TxEvent } from '@webb-tools/abstract-api-provider';
 import Button from '@webb-tools/webb-ui-components/components/buttons/Button';
 import { useCallback, useMemo, useState } from 'react';
-
 import {
   type CancelWithdrawRequestContext,
   type ExecuteAllWithdrawRequestContext,
-  TxEvent,
 } from '../../../data/restake/RestakeTx/base';
 import useRestakeTx from '../../../data/restake/useRestakeTx';
 import useRestakeTxEventHandlersWithNoti from '../../../data/restake/useRestakeTxEventHandlersWithNoti';
