+++ libs/abstract-api-provider/src/index.ts
@@ -3,3 +3,4 @@
 
 export * from './account';
 export * from './webb-provider.interface';
+export * from './transaction-events';
